feat(mm): Implement Transparent Huge Page (THP) promotion and management
- Added functions to find THP candidates and promote them to huge pages in `ProcessAddressSpace`.
- Introduced `ThpCandidate` struct to represent potential THP regions.
- Enhanced `ThpPromotionManager` to scan for candidates and promote them based on usage.
- Updated memory allocation logic to support huge frame allocation in `buddy_allocator`.
- Integrated THP promotion logic into the existing memory management framework.
- Added a minimal NVMe shim for task-scoped IO operations.
- Refactored RCU state management to be per-CPU, improving performance and scalability.
- Introduced new methods for handling RCU callbacks and synchronization.
- Updated tests to cover new THP functionality and NVMe operations.

+++ kernel/src/mm/tlb_batch.rs
@@ -615,7 +615,7 @@ pub fn flush_tlb_all() {
 /// 単一CPUにTLBフラッシュIPIを送信
 fn send_tlb_ipi_to_cpu(cpu_id: usize) {
     // CPU IDをAPIC IDとして使用（通常は1:1マッピング）
-    let apic_id = cpu_id as u8;
+    let apic_id = cpu_id as u32;
     
     // interrupt_manager経由でIPI送信
     crate::io::interrupt_manager::send_ipi(apic_id, TLB_FLUSH_VECTOR);
